+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - [crawler] Enable usage of the GitHub user PRs crawler via the Monocle config.
+- [cli] Add a Janitor command to reset the commit date of a crawler
 
 ### Changed